Python/pyproject.toml
[project]
name = "unreal-mcp"
version = "0.1.0"
description = "Unreal MCP Server: A server for Unreal Engine integration via the Model Context Protocol (MCP)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.4.1",
  "fastmcp>=0.2.0",
  "uvicorn>=0.30.0",
  "starlette>=0.38.0",
  "fastapi",
  "pydantic>=2.10.0",
  "requests",
  "httpx>=0.25.0",
  "python-dotenv>=1.0.0",
  "Pillow>=10.0.0"
]

[project.optional-dependencies]
test = [
  "pytest>=7.0.0",
  "pytest-html>=4.0.0",
  "pytest-timeout>=2.0.0"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# The main server script is a single-file module
py-modules = ["unreal_mcp_server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "umg: UMG Widget操作テスト",
    "blueprint: Blueprint操作テスト",
    "node: Blueprintノード操作テスト",
    "gas: GAS操作テスト",
    "integration: 統合テスト",
    "slow: 遅いテスト"
]
addopts = "-v --tb=short"
